+++ editor/src/debugger/editor-client.ts
@@ -1,150 +1,107 @@
-import { 
-  debugLogger, 
-  DebuggerEvent, 
+import {
+  DebuggerEvent,
+  DebuggerEventType,
+  ERROR_PIN_ID,
+  PinDebuggerEvent,
   PinType,
-  RemoteDebuggerCallback,
-  RemoteDebuggerCancelFn,
-  enumToArray,
-  DebuggerServerEventType,
-  HistoryPayload
 } from "@flyde/core";
+import { HistoryPayload } from "@site/../remote-debugger/dist";
 
-import { default as axios } from "axios";
+const MAX_EVENTS_HISTORY_LIMIT = 200;
 
-import { io as _io } from "socket.io-client";
+type DebugHistoryMap = Map<
+  string,
+  { total: number; lastSamples: DebuggerEvent[] }
+>;
 
-const debug = debugLogger("remote-debugger:editor-client");
-
-export type GetPinHistoryDto = {
-  id?: string;
+export type HistoryPlayer = {
+  addEvents: (events: DebuggerEvent[]) => void;
+  requestHistory: (
+    insId: string,
+    pinId: string,
+    type: PinType
+  ) => Promise<HistoryPayload>;
+  clear: () => void;
 };
 
-export type GetHistoryDto = {
-  pinId?: string;
-  type?: PinType;
-  insId: string;
-  limit?: number;
-  executionId: string;
-};
-
-export type EditorDebuggerClient = {
-  emitChange: (data: {}) => void;
-  emitBreakpointsChange: (insIdsAndPins: string[]) => void;
-
-  onBatchedEvents: (
-    cb: RemoteDebuggerCallback<DebuggerEvent[]>
-  ) => RemoteDebuggerCancelFn;
-
-  onRuntimeReady: (cb: RemoteDebuggerCallback<void>) => RemoteDebuggerCancelFn;
-  interceptInput: (value: DebuggerEvent) => void;
-  interceptOutput: (value: DebuggerEvent) => void;
-  onChangeAwk: (
-    cb: RemoteDebuggerCallback<{ hash: string }>
-  ) => RemoteDebuggerCancelFn;
-  onChangeError: (cb: (error: any) => void) => RemoteDebuggerCancelFn;
-  onIsAlive: (
-    cb: RemoteDebuggerCallback<{ time: number }>
-  ) => RemoteDebuggerCancelFn;
-
-  emitInputValue: (pinId: string, value: any) => void;
-
-  destroy: () => void;
-  onDisconnect: (cb: () => void) => void;
-  debugInfo(): string;
-  requestState: () => Promise<any>;
-
-  getHistory: (dto: GetHistoryDto) => Promise<HistoryPayload>;
-  clearHistory: () => Promise<void>;
-
-  triggerNode: (nodeId: string, inputs: Record<string, any>) => void;
-};
-
-export const createEditorClient = (
-  url: string,
-  executionId: string
-): EditorDebuggerClient => {
-  const urlNodes = new URL(url);
-
-  const socket = _io(urlNodes.origin, {
-    path: `${
-      urlNodes.pathname === "/" ? "" : urlNodes.pathname
-    }/socket.io/editor`,
-    timeout: 30000,
-  });
-
-  socket.emit("join-room-editor", executionId);
+export const createHistoryPlayer = (): HistoryPlayer => {
+  const pinHistoryMap: DebugHistoryMap = new Map();
+  const insHistoryMap: DebugHistoryMap = new Map();
 
   return {
-    emitChange: (data) => {
-      socket.emit(DebuggerServerEventType.CHANGE_EVENT_NAME, data);
-    },
-    emitBreakpointsChange: (data) => {
-      socket.emit(DebuggerServerEventType.UPDATE_BREAKPOINTS, data);
-    },
-    interceptInput: (data: DebuggerEvent) => {
-      socket.emit(DebuggerServerEventType.INPUT_VALUE_OVERRIDE, data);
-    },
-    interceptOutput: (data: DebuggerEvent) => {
-      socket.emit(DebuggerServerEventType.OUTPUT_VALUE_OVERRIDE, data);
-    },
-    onRuntimeReady: (cb) => {
-      socket.on(DebuggerServerEventType.RUNTIME_READY, cb);
-      return () => socket.off(DebuggerServerEventType.RUNTIME_READY, cb);
-    },
-    onChangeAwk: (cb) => {
-      socket.on(DebuggerServerEventType.CHANGE_AWK, cb);
-      return () => socket.off(DebuggerServerEventType.CHANGE_AWK, cb);
-    },
-    onChangeError: (cb) => {
-      socket.on(DebuggerServerEventType.CHANGE_ERROR, cb);
-      return () => socket.off(DebuggerServerEventType.CHANGE_ERROR, cb);
-    },
-    onIsAlive: (cb) => {
-      socket.on(DebuggerServerEventType.IS_ALIVE, cb);
-      return () => socket.off(DebuggerServerEventType.IS_ALIVE, cb);
-    },
-    emitInputValue: (pinId, value) => {
-      debug(`Emitting push input value to ${pinId} %o`, value);
-      socket.emit(DebuggerServerEventType.PUSH_INPUT_VALUE, { pinId, value });
-    },
-    destroy: () => {
-      socket.disconnect();
+    requestHistory: async (insId: string, pinId: string, pinType: PinType) => {
+      const type =
+        pinType === "input"
+          ? DebuggerEventType.INPUT_CHANGE
+          : DebuggerEventType.OUTPUT_CHANGE;
 
-      enumToArray(DebuggerServerEventType).forEach((type) => socket.off(type));
+      if (pinId) {
+        const id = `${insId}.${pinId}.${type}`;
+        const payload: HistoryPayload = pinHistoryMap.get(id) || {
+          total: 0,
+          lastSamples: [],
+        };
+        const samples = payload.lastSamples.slice(0, MAX_EVENTS_HISTORY_LIMIT);
+        return { ...payload, lastSamples: samples };
+      } else {
+        const payload: HistoryPayload = insHistoryMap.get(insId) || {
+          total: 0,
+          lastSamples: [],
+        };
+        const samples = payload.lastSamples.slice(0, MAX_EVENTS_HISTORY_LIMIT);
+        return { ...payload, lastSamples: samples };
+      }
     },
-    onDisconnect: (cb) => {
-      socket.on("disconnect", cb);
-      return () => socket.off("disconnect", cb);
+    addEvents: (events) => {
+      events.forEach((event) => {
+        // hack for easily logging errors to error pin
+        if (event.type === DebuggerEventType.ERROR) {
+          const ev: PinDebuggerEvent<DebuggerEventType.OUTPUT_CHANGE> =
+            event as any;
+          ev.type = DebuggerEventType.OUTPUT_CHANGE;
+          ev.pinId = ERROR_PIN_ID;
+        }
+
+        if (
+          event.type === DebuggerEventType.INPUT_CHANGE ||
+          event.type === DebuggerEventType.OUTPUT_CHANGE
+        ) {
+          const { insId, type, pinId } = event;
+          {
+            const id = `${insId}.${pinId}.${type}`;
+            const curr = pinHistoryMap.get(id) || { total: 0, lastSamples: [] };
+            curr.lastSamples.unshift(event);
+            if (curr.lastSamples.length > MAX_EVENTS_HISTORY_LIMIT) {
+              curr.lastSamples.splice(
+                MAX_EVENTS_HISTORY_LIMIT,
+                curr.lastSamples.length - MAX_EVENTS_HISTORY_LIMIT
+              );
+            }
+            curr.total++;
+            pinHistoryMap.set(id, curr);
+          }
+
+          {
+            const curr = insHistoryMap.get(insId) || {
+              total: 0,
+              lastSamples: [],
+            };
+            curr.lastSamples.unshift(event);
+            if (curr.lastSamples.length > MAX_EVENTS_HISTORY_LIMIT) {
+              curr.lastSamples.splice(
+                MAX_EVENTS_HISTORY_LIMIT,
+                curr.lastSamples.length - MAX_EVENTS_HISTORY_LIMIT
+              );
+            }
+            curr.total++;
+            pinHistoryMap.set(insId, curr);
+          }
+        }
+      });
     },
-    debugInfo: () => {
-      return `Remote debugger for ${url}`;
-    },
-    onBatchedEvents: (cb) => {
-      socket.on(DebuggerServerEventType.EVENTS_BATCH, cb);
-      return () => socket.off(DebuggerServerEventType.EVENTS_BATCH, cb);
-    },
-    requestState: () => {
-      return axios.get(`${url}/state`).then((r: any) => r.data.state);
-    },
-    getHistory: (dto) => {
-      return axios
-        .get(`${url}/history`, {
-          params: {
-            insId: dto.insId,
-            pinId: dto.pinId,
-            limit: dto.limit,
-            executionId,
-          },
-        })
-        .then((r: any) => r.data);
-    },
-    clearHistory: () => {
-      return axios.delete(`${url}/history`).then(() => {});
-    },
-    triggerNode: (nodeId, inputs) => {
-      return axios
-        .post(`${url}/trigger`, { nodeId, inputs })
-        .then((r: any) => r.data);
+    clear: () => {
+      pinHistoryMap.clear();
+      insHistoryMap.clear();
     },
   };
 };
