+++ editor/src/index-no-tailwind.scss
@@ -1,0 +1,6 @@
+@import url("https://fonts.googleapis.com/css2?family=DM+Sans:wght@300;400;500;600;700");
+
+// Import component styles only (no Tailwind)
+@import "./flow-editor/style.scss";
+@import "./visual-node-editor/style.scss";
+@import "./lib/react-utils/react-resizable.scss";
