--- website/app/globals.css
@@ -1,65 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-:root {
-  --primary: #2887F4;
-  --background: #ffffff;
-  --foreground: #171717;
-  --card: #f9f9f9;
-  --card-foreground: #171717;
-  --border: #e5e5e5;
-  --muted: #f1f1f1;
-  --muted-foreground: #737373;
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --primary: #2887F4;
-    --background: #0a0a0a;
-    --foreground: #ededed;
-    --card: #171717;
-    --card-foreground: #ededed;
-    --border: #333333;
-    --muted: #262626;
-    --muted-foreground: #a3a3a3;
-  }
-}
-
-body {
-  background: var(--background);
-  color: var(--foreground);
-  font-family: var(--font-sans);
-}
-
-.flyde-flow-editor {
-  width: 100%;
-  height: 100%;
-  /* overflow: hidden; */
-  position: relative;
-}
-
-.canvas-positioning-hack {
-  /* display: flex; */
-  width: 100%;
-  height: 100%;
-}
-
-/* flyde hacks */
-/* cmd +k */
-.flyde-flow-editor .absolute.top-4.right-5.z-10 {
-  display: none;
-}
-
-.flyde-flow-editor .help-bubble {
-  display: none;
-}
-
-/* hide onboarding tips */
-.flyde-flow-editor .onboarding-tips {
-  display: none;
-}
-
-.flyde-flow-editor {
-  /* pointer-events: none; */
-}
