+++ website/scripts/sync-examples.ts
@@ -1,11 +1,8 @@
 import * as resolver from '@flyde/loader';
 import { join } from 'path';
-import { writeFileSync, mkdirSync } from 'fs';
+import { writeFileSync, mkdirSync, readdirSync } from 'fs';
 
-const examples = [
-    'ExampleBlogpost',
-    'ExampleChatbot'
-];
+const examples = readdirSync(join(__dirname, '../flyde')).filter(file => file.endsWith('.flyde')).map(file => file.replace('.flyde', ''));
 
 // Create resolved directory if it doesn't exist
 const resolvedDir = join(__dirname, '../flyde/resolved');
@@ -17,7 +14,7 @@ examples.forEach(exampleName => {
     const flow = resolver.resolveFlowByPath(flowPath);
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const editorNode = resolver.resolveEditorNode(flow as any, resolver.createServerReferencedNodeFinder(flowPath));
-    
+
     // Write to TypeScript file
     const outputPath = join(resolvedDir, `${exampleName}.ts`);
     const content = `export const ${exampleName} = ${JSON.stringify(editorNode, null, 2)};`;
